using the prepare lifecycle script
This commit uses the `prepare` lifecycle script in order to make sure
the `build` step of each workspace is run prior to trying to link the
bin files for each of these.

An unfortunate requirement here is that packages that need to be have
a build step before they can be used in other workspaces might need to
be ordered in the file system, that's the reason I moved `packages/sdk`
to `core/sdk`.